chore: code clean up and verbose logging
