+++ src/run.ts
@@ -1,8 +1,8 @@
 import { Octokit } from "@octokit/rest";
-import { updateTasks } from "./helpers/update-tasks";
 import { Context } from "./types/context";
 import { PluginInputs } from "./types/plugin-inputs";
 import { Logs } from "@ubiquity-dao/ubiquibot-logger";
+import { watchUserActivity } from "./handlers/watch-user-activity";
 
 export async function run(inputs: PluginInputs) {
   const octokit = new Octokit({ auth: inputs.authToken });
@@ -11,12 +11,12 @@ export async function run(inputs: PluginInputs) {
     payload: inputs.eventPayload,
     config: inputs.settings,
     octokit,
-    logger: new Logs("info"),
+    logger: new Logs("verbose"),
   };
   await runPlugin(context);
   return JSON.stringify({ status: 200 });
 }
 
 export async function runPlugin(context: Context) {
-  return await updateTasks(context);
+  return await watchUserActivity(context);
 }
